Added optional translation search plus a few usability improvements
The best ins file is copied to name_orient.ins, the refinement is run
once more, and the dummy atoms are deleted from the res file.

Main reads an optional translation size tmax from the command line, and
passes it to proc_file. If so, the dummy atoms are shifted a random
distance up to tmax in a random direction before orienting the fragment.

The readme file was updated for readability.